bunch of changes from @matrixagent
 - Removed/updated some legacy code that would cause warnings in the
logfile
 - The site (e.g. DeviceBondage or Hogtied) is now set as the "movie
studio", previously it would always just be "Kink.com". This allows for
awesome sorting/grouping by site.
 - The title is no longer in the format "$sitename - $episodeid", but now
actually also retrieved from the shoot info in the website. The episode
ID is still appended in brackets for easy reference. Additionally, the
tagline is set to "sitename - $episodeid".
 - Instead of just fetching one of the images as the movie poster, it now
is a] fetching all available images, and b] fetches them in the highest
quality possible. The one image that the shoot site is using as their
primary/showcase picture is set as the poster, the other images are set
as "art", so they'll be used as backdrops.
 - Genres were completely removed as there are none stated on the website.